move getScmServiceId to HddsUtils
